First Commit
This will probably be the biggest commit for the site! This commit adds the home page, Smash results, projects, content, and other section. Smash results is split into Notable Wins, Offline Results, and Online Results. There is one project added to the projects section which is the Hiking Project. The full project will be added in later. The content sections are split into Graphic Design, Video Editing, and Game Modding. The Video Editing section is also split into jokes, Smash, and other. Smash results, projects, and content will be updated as time goes on.